requirements.txt
# Core dependencies
requests>=2.28.0
SQLAlchemy>=2.0.0
python-dotenv>=0.19.0

# Amazon Integration
boto3>=1.26.0
python-amazon-sp-api>=0.18.0

# Caching and Performance
redis>=4.3.0
aioredis>=2.0.0

# API and Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Testing and Development
pytest>=7.0.0
requests-mock>=1.9.0

# Monitoring and Logging
prometheus-client>=0.12.0
python-json-logger>=2.0.0